Renames gem to sony_ci_api
We didn't have perms to push to rubygems.org because it was originally set up by
Chuck, and when asked, he couldn't remember the password, and he apparently used
his wgbh.org email, to which he no longer has access.

So we opted to rename it. Also, the new name adheres closer to ruby gem naming
conventions.

Closes #13.

Also...

* Fixes an test checking for the download URL of uploaded things. This appears
  to have changed from an S3 URL to a Cloudfront URL.
* Changes mocked request to fetch Oauth2 token. Previously, it was prepending
  the URL encoded username and password to the URL, but the call is being made
  without it now, instead passing credentials as POST params. I'm guessing this
  was causing test failures earlier, before this commit. However, they are
  flagged to not run on Travis, so failures may have been overlooked.
* Removes sony_ci_api.rb at the gem's root. This appears to be redundant with
  bin/sony_ci_api, and I couldn't find any references to it.
* Removes Gemfile.lock from revision.